Code restructuring
(1) Create ActivityDashboard.js component and move all code
related to Activity feature from App.js.
(2) Create Activity.css and move all styling
related to Activity feature from App.css.
(3) Create UserInfoDashboard.js component and move all code
related to Activity feature from App.js.
(4) Create UserInfo.css and move all styling
related to Activity feature from App.css.
(5) Add <ActivityDashboard /> and <UserInfoDashboard /> to App.js
(6) All original functions still work just fine
