+++ src/css/App.css
@@ -7,77 +7,3 @@ html {
   font-size: 62.5%; /* 1 rem = 10px;*/
   padding: 2rem;
 }
-
-fieldset {
-  align-items: flex-start;
-  display: flex;
-  flex-direction: column;
-  font-size: 2rem;
-  gap: 1rem;
-  justify-content: center;
-  padding: 1rem;
-}
-
-legend {
-  text-align: center;
-}
-
-.App {
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-}
-
-input,
-select {
-  font-size: 1.75rem;
-  margin-left: 1rem;
-}
-
-select {
-  padding: 0.2rem;
-}
-
-.user-info-form {
-  display: flex;
-  flex-direction: column;
-  width: 40%;
-}
-
-.activity-form {
-  display: flex;
-  flex-direction: column;
-  width: 40%;
-}
-
-.submit-button {
-  align-self: flex-end;
-  padding: 0.5rem;
-}
-
-.activity-item {
-  display: flex;
-  flex-direction: column;
-  font-size: 2rem;
-  border: 0.1rem solid;
-  width: 20%;
-  padding: 1rem;
-  position: relative;
-}
-
-.user-info {
-  display: flex;
-  flex-direction: column;
-  font-size: 2rem;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 0.5rem;
-}
-
-.delete-button {
-  position: absolute;
-  right: 1rem;
-  top: 1rem;
-  font-size: 1.5rem;
-  border: none;
-}
