+++ src/css/UserInfo.css
@@ -1,0 +1,18 @@
+fieldset {
+  align-items: flex-start;
+  display: flex;
+  flex-direction: column;
+  font-size: 2rem;
+  gap: 1rem;
+  justify-content: center;
+  padding: 1rem;
+}
+
+legend {
+  text-align: center;
+}
+
+input {
+  font-size: 1.75rem;
+  margin-left: 1rem;
+}
